Fix DB-related test issues
