+++ tests/test_orchestrator.py
@@ -1,12 +1,7 @@
 import pytest
 
 from smartsim import Experiment
-from smartsim.database import (
-    CobaltOrchestrator,
-    Orchestrator,
-    PBSOrchestrator,
-    SlurmOrchestrator,
-)
+from smartsim.database import Orchestrator
 from smartsim.error import SmartSimError
 from smartsim.error.errors import SSUnsupportedError
 
@@ -86,7 +81,7 @@ def test_catch_local_db_errors():
 
 
 def test_pbs_set_run_arg():
-    orc = PBSOrchestrator(6780, db_nodes=3, batch=False, interface="lo")
+    orc = Orchestrator(6780, db_nodes=3, batch=False, interface="lo", launcher="pbs", run_command="aprun")
     orc.set_run_arg("account", "ACCOUNT")
     assert all(
         [db.run_settings.run_args["account"] == "ACCOUNT" for db in orc.entities]
@@ -98,11 +93,11 @@ def test_pbs_set_run_arg():
 
 
 def test_pbs_set_batch_arg():
-    orc = PBSOrchestrator(6780, db_nodes=3, batch=False, interface="lo")
+    orc = Orchestrator(6780, db_nodes=3, batch=False, interface="lo", launcher="pbs", run_command="aprun")
     with pytest.raises(SmartSimError):
         orc.set_batch_arg("account", "ACCOUNT")
 
-    orc2 = PBSOrchestrator(6780, db_nodes=3, batch=True, interface="lo")
+    orc2 = Orchestrator(6780, db_nodes=3, batch=True, interface="lo", launcher="pbs", run_command="aprun")
     orc2.set_batch_arg("account", "ACCOUNT")
     assert orc2.batch_settings.batch_args["account"] == "ACCOUNT"
     orc2.set_batch_arg("N", "another_name")
@@ -113,7 +108,7 @@ def test_pbs_set_batch_arg():
 
 
 def test_slurm_set_run_arg():
-    orc = SlurmOrchestrator(6780, db_nodes=3, batch=False, interface="lo")
+    orc = Orchestrator(6780, db_nodes=3, batch=False, interface="lo", launcher="slurm")
     orc.set_run_arg("account", "ACCOUNT")
     assert all(
         [db.run_settings.run_args["account"] == "ACCOUNT" for db in orc.entities]
@@ -121,11 +116,11 @@ def test_slurm_set_run_arg():
 
 
 def test_slurm_set_batch_arg():
-    orc = SlurmOrchestrator(6780, db_nodes=3, batch=False, interface="lo")
+    orc = Orchestrator(6780, db_nodes=3, batch=False, interface="lo", launcher="slurm")
     with pytest.raises(SmartSimError):
         orc.set_batch_arg("account", "ACCOUNT")
 
-    orc2 = SlurmOrchestrator(6780, db_nodes=3, batch=True, interface="lo")
+    orc2 = Orchestrator(6780, db_nodes=3, batch=True, interface="lo", launcher="slurm")
     orc2.set_batch_arg("account", "ACCOUNT")
     assert orc2.batch_settings.batch_args["account"] == "ACCOUNT"
 
@@ -133,8 +128,8 @@ def test_slurm_set_batch_arg():
 ###### Cobalt ######
 
 
-def test_set_run_arg():
-    orc = CobaltOrchestrator(6780, db_nodes=3, batch=False, interface="lo")
+def test_cobalt_set_run_arg():
+    orc = Orchestrator(6780, db_nodes=3, batch=False, interface="lo", launcher="cobalt", run_command="aprun")
     orc.set_run_arg("account", "ACCOUNT")
     assert all(
         [db.run_settings.run_args["account"] == "ACCOUNT" for db in orc.entities]
@@ -145,13 +140,59 @@ def test_set_run_arg():
     )
 
 
-def test_set_batch_arg():
-    orc = CobaltOrchestrator(6780, db_nodes=3, batch=False, interface="lo")
+def test_cobalt_set_batch_arg():
+    orc = Orchestrator(6780, db_nodes=3, batch=False, interface="lo", launcher="cobalt", run_command="aprun")
     with pytest.raises(SmartSimError):
         orc.set_batch_arg("account", "ACCOUNT")
 
-    orc2 = CobaltOrchestrator(6780, db_nodes=3, batch=True, interface="lo")
+    orc2 = Orchestrator(6780, db_nodes=3, batch=True, interface="lo", launcher="cobalt", run_command="aprun")
     orc2.set_batch_arg("account", "ACCOUNT")
     assert orc2.batch_settings.batch_args["account"] == "ACCOUNT"
     orc2.set_batch_arg("outputprefix", "new_output/")
     assert "outputprefix" not in orc2.batch_settings.batch_args
+
+
+
+###### LSF ######
+
+def test_catch_orc_errors_lsf():
+    with pytest.raises(SSUnsupportedError):
+        orc = Orchestrator(6780, db_nodes=2, db_per_host=2, batch=False, launcher="lsf")
+
+    orc = Orchestrator(
+        6780,
+        db_nodes=3,
+        batch=False,
+        hosts=["batch", "host1", "host2"],
+        launcher="lsf"
+    )
+    with pytest.raises(SmartSimError):
+        orc.set_batch_arg("P", "MYPROJECT")
+
+
+def test_lsf_set_run_args():
+
+    orc = Orchestrator(
+        6780,
+        db_nodes=3,
+        batch=True,
+        hosts=["batch", "host1", "host2"],
+        launcher="lsf"
+    )
+    orc.set_run_arg("l", "gpu-gpu")
+    assert all(["l" not in db.run_settings.run_args for db in orc.entities])
+
+
+def test_lsf_set_batch_args():
+
+    orc = Orchestrator(
+        6780,
+        db_nodes=3,
+        batch=True,
+        hosts=["batch", "host1", "host2"],
+        launcher="lsf"
+    )
+
+    assert orc.batch_settings.batch_args["m"] == '"batch host1 host2"'
+    orc.set_batch_arg("D", "102400000")
+    assert orc.batch_settings.batch_args["D"] == "102400000"
